year_3/comppi_0/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-background-tasks==1.1.12
django-common-helpers==0.9.1
django-compat==1.0.15
django-cron==0.5.0
future==0.16.0
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
python-telegram-bot==9.0.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
